+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Graceful shutdown of event handlers ([#431](https://github.com/commanded/commanded/pull/431)).
 - Ensure command dispatch metadata is a map ([#432](https://github.com/commanded/commanded/pull/432)).
 - Retry command execution on node down ([#429](https://github.com/commanded/commanded/pull/429)).
+- Dispatch returning resultant events ([#444](https://github.com/commanded/commanded/pull/444)).
 
 ## v1.2.0